% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_name.R
\name{clean_name}
\alias{clean_name}
\title{Clean a string from characters unwanted in (mostly) yishuv names}
\usage{
clean_name(name)
}
\arguments{
\item{name}{a character vector}
}
\value{
a character vector without leading and trailing white space, and with
characters that are only letters and digits, excluding the following: \verb{'-()\\"}
}
\description{
Clean a string from characters unwanted in (mostly) yishuv names
}
\examples{
clean_name("test-123_test*456")
x <- read_cbs_yishuv(system.file("extdata", "bycode2021.xlsx",
 package = "il.cbs.muni"))[[1]][c(153, 342)]
x
clean_name(x)
}
