% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/il.cbs.muni-package.R
\docType{package}
\name{il.cbs.muni-package}
\alias{il.cbs.muni}
\alias{il.cbs.muni-package}
\title{il.cbs.muni: Utility Functions to Work with Israeli Central Bureau of Statistics Municipal Data}
\description{
Analyst oriented utility functions to handle the different quirks of the Israeli CBS municipal data, harmonize id's and bring together data points from different years.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matanhakim/il.cbs.muni}
  \item \url{https://matanhakim.github.io/il.cbs.muni/}
  \item Report bugs at \url{https://github.com/matanhakim/il.cbs.muni/issues}
}

}
\author{
\strong{Maintainer}: Matan Hakim \email{matanhakim@gmail.com} (\href{https://orcid.org/0009-0002-9372-8975}{ORCID}) [copyright holder]

}
\keyword{internal}
