% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_yishuv_id.R
\name{pad_yishuv_id}
\alias{pad_yishuv_id}
\title{Pad a yishuv id to a 4-digit character vector}
\usage{
pad_yishuv_id(yishuv_id)
}
\arguments{
\item{yishuv_id}{a character vector containing only characters and numbers, where each
element is no longer than 4 characters or digits.}
}
\value{
A character vector the same length as \code{yishuv_id}, where each element is
4 characters long and left-padded with \code{"0"}.
}
\description{
Pad a yishuv id to a 4-digit character vector
}
\examples{
x <- c(1, "23", "4000", 5600)
pad_yishuv_id(x)
}
