% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cbs_muni.R
\name{read_cbs_muni}
\alias{read_cbs_muni}
\title{Read a municipal data file to a tibble}
\usage{
read_cbs_muni(path, year, data_domain = c("physical", "budget"))
}
\arguments{
\item{year}{A numeric vector of length 1 denoting which year the data file
pointed in \code{path} is for. Currently supporting only 2016 and later, since before
2016 regional councils are split from cities and local councils.}

\item{data_domain}{A character vector of length 1, one of \code{c("physical", "budget")}.
Every Excel municipal data file has a few different data domains, most notably
physical and population data, and budget data.}
}
\value{
A tibble with municipal data for a specific year, where every row is a
municipality and every column is a different variable for this municipality in
that year. Be advised all columns are of type character, so you nee to parse
the data types yourself at will.
}
\description{
This function is a wrapper around \code{readxl::readexcel()}, reading a specific
municipal data file for a specific year and a specific data domain. Its added
value is in its use of \code{row_to_names_fill()} and its pre-defined parameters for
every year and its specific quirks in the Excel headers.
}
\examples{
df <- read_cbs_muni(
  system.file("extdata", "p_libud_2021.xlsx", package = "il.cbs.muni"),
  year = 2021,
  data_domain = "physical"
)

df |>
  dplyr::select(1:15) |>
  dplyr::glimpse()
}
