% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_muni_id.R
\name{modify_muni_id}
\alias{modify_muni_id}
\title{Modify a municipal id depending on the municipal status and yishuv id}
\usage{
modify_muni_id(muni_id, yishuv_id)
}
\arguments{
\item{muni_id}{a character or numeric vector indicating the municipal status,
where "0" or "99" represents a city or a local council, and every other two-digit
number or character represents a regional council.}

\item{yishuv_id}{a character or numeric vector representing the yishuv id.
should be 4 digits long according to the il.verse conventions.}
}
\value{
a character vector with 4 digits municipal id for cities and local councils
and 2 digits municipal id for regional councils.
}
\description{
Modify a municipal id depending on the municipal status and yishuv id
}
\examples{
muni_id <- c(0, 99, 1, 2)
yishuv_id <- c("0001", "1000", "1234", "1567")
modify_muni_id(muni_id, yishuv_id)
}
