% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cbs_muni.R, R/read_cbs_yishuv.R
\name{read_cbs_muni}
\alias{read_cbs_muni}
\title{Read a municipal data file to a tibble}
\usage{
read_cbs_muni(path, cols = NULL)

read_cbs_muni(path, cols = NULL)
}
\arguments{
\item{path}{A character vector of length 1, denoting the local file path to the
yishuvim data file. A full list of available files by the CBS is at the
\href{https://www.cbs.gov.il/he/publications/Pages/2019/\%D7\%99\%D7\%99\%D7\%A9\%D7\%95\%D7\%91\%D7\%99\%D7\%9D-\%D7\%91\%D7\%99\%D7\%A9\%D7\%A8\%D7\%90\%D7\%9C.aspx}{relevant CBS page}.}

\item{cols}{<\href{https://dplyr.tidyverse.org/reference/dplyr_tidy_select.html}{tidy-select}>
Columns to keep. The default \code{NULL} keeps all columns.}

\item{year}{A numeric vector of length 1 denoting which year the data file
pointed in \code{path} is for. Currently supporting only 2003 and later, since before
2003 the data structure is very different.}

\item{muni_type}{A character vector of length 1, one of
\code{c("all", "city_lc", "rc")}. Since 2016, all municipal types are bundled together
in the same sheets, but before 2016 there are different sheets for cities and
local councils (\code{"city_lc"}) and regional councils (\code{"rc"}). This parameter
chooses which sheet you would read.}

\item{data_domain}{A character vector of length 1, one of
\code{c("physical", "budget", "summary", "labor_force_survey", "social_survey")}.
Every Excel municipal data file has a few different data domains, most notably
physical and population data, and budget data.}
}
\value{
A tibble with municipal data for a specific year, where every row is a
municipality and every column is a different variable for this municipality in
that year. Be advised all columns are of type character, so you need to parse
the data types yourself at will. Column names are merged from the relevant headers,
and only single whitespaces are kept. Rows with more than 90\% empty cells (usually
rows with non-data notes) are removed.

A tibble with yishuvim data for a specific year, where every row is a
yishuv and every column is a different variable for this yishuv in
that year. Be advised all columns are of type character, so you need to parse
the data types yourself at will. Column names are cleaned so only single
whitespaces are kept.
}
\description{
This function is a wrapper around \code{readxl::read_excel()}, reading a specific
municipal data file for a specific year and a specific data domain. Its added
value is in its use of \code{row_to_names_fill()} and its pre-defined parameters for
every year and its specific quirks in the Excel headers. For advanced users,
the full set of options is available with \code{il.cbs.muni:::df_cbs_muni_params}.

This function is a wrapper around \code{readxl::read_excel()}, reading a specific
yishuvim data file or a part of it. A yishuv, or a point of residence, is a
geographically defined place where people live. Some yishuvim are municipalities,
in the case of of cities and local councils, but most are not. most yishuvim are
part of municipalities that are regional councils. Also, some yishuvim are not
themselves and are not part of a municipality, like some Bedouin places in
southern Israel, some industry areas, Mikveh Israel, and more.
}
\examples{
df <- read_cbs_muni(
  system.file("extdata", "p_libud_2021.xlsx", package = "il.cbs.muni"),
  year = 2021,
  data_domain = "physical"
)

df |>
  dplyr::select(1:15) |>
  dplyr::glimpse()
library(dplyr)
read_cbs_yishuv(system.file("extdata", "bycode2021.xlsx", package = "il.cbs.muni")) |>
  dplyr::glimpse()

read_cbs_yishuv(
  system.file("extdata", "bycode2021.xlsx", package = "il.cbs.muni"),
  cols = c(1, 2, 5, 13)
) |>
  mutate(across(2, pad_yishuv_id)) |>
  glimpse()
}
